+++ packages/editor/editor-json-transformer/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @atlaskit/editor-json-transformer
 
+## 2.5.11
+
 ## 2.5.10
 
 ## 2.5.9