+++ packages/editor/editor-markdown-transformer/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @atlaskit/editor-markdown-transformer
 
+## 0.2.4
+
 ## 0.2.3
 
 ## 0.2.2